+++ src/client.rs
@@ -790,13 +790,6 @@ pub enum ClientSet {
     /// Systems that reset the client.
     ///
     /// Runs in [`PreUpdate`] when the client just disconnected.
-    ///
-    /// You may want to disable this set if you want to preserve client replication state across reconnects.
-    /// In that case, you need to manually repair the client state (or use something like
-    /// [`bevy_replicon_repair`](https://docs.rs/bevy_replicon_repair)).
-    ///
-    /// If this set is disabled and you don't want to repair client state, then you need to manually clean up
-    /// the client after a disconnect or when reconnecting.
     Reset,
 }
 
@@ -810,13 +803,11 @@ pub enum ClientSet {
 pub struct ServerUpdateTick(RepliconTick);
 
 /// Cached buffered mutate messages, used to synchronize mutations with update messages.
-///
-/// If [`ClientSet::Reset`] is disabled, then this needs to be cleaned up manually with [`Self::clear`].
 #[derive(Default, Resource)]
-pub struct BufferedMutations(Vec<BufferedMutate>);
+pub(crate) struct BufferedMutations(Vec<BufferedMutate>);
 
 impl BufferedMutations {
-    pub fn clear(&mut self) {
+    fn clear(&mut self) {
         self.0.clear();
     }