+++ tests/mutations.rs
@@ -409,10 +409,10 @@ fn marker() {
 
     let client_entity = client_app.world_mut().spawn(ReplaceMarker).id();
 
-    let test_client_entity = **client_app.world().resource::<TestClientEntity>();
+    let client = **client_app.world().resource::<TestClientEntity>();
     let mut entity_map = server_app
         .world_mut()
-        .get_mut::<ClientEntityMap>(test_client_entity)
+        .get_mut::<ClientEntityMap>(client)
         .unwrap();
     entity_map.insert(server_entity, client_entity);
 
@@ -472,10 +472,10 @@ fn marker_with_history() {
 
     let client_entity = client_app.world_mut().spawn(HistoryMarker).id();
 
-    let test_client_entity = **client_app.world().resource::<TestClientEntity>();
+    let client = **client_app.world().resource::<TestClientEntity>();
     let mut entity_map = server_app
         .world_mut()
-        .get_mut::<ClientEntityMap>(test_client_entity)
+        .get_mut::<ClientEntityMap>(client)
         .unwrap();
     entity_map.insert(server_entity, client_entity);
 
@@ -556,10 +556,10 @@ fn marker_with_history_consume() {
 
     let client_entity = client_app.world_mut().spawn(HistoryMarker).id();
 
-    let test_client_entity = **client_app.world().resource::<TestClientEntity>();
+    let client = **client_app.world().resource::<TestClientEntity>();
     let mut entity_map = server_app
         .world_mut()
-        .get_mut::<ClientEntityMap>(test_client_entity)
+        .get_mut::<ClientEntityMap>(client)
         .unwrap();
     entity_map.insert(server_entity, client_entity);
 
@@ -638,10 +638,10 @@ fn marker_with_history_old_update() {
 
     let client_entity = client_app.world_mut().spawn(HistoryMarker).id();
 
-    let test_client_entity = **client_app.world().resource::<TestClientEntity>();
+    let client = **client_app.world().resource::<TestClientEntity>();
     let mut entity_map = server_app
         .world_mut()
-        .get_mut::<ClientEntityMap>(test_client_entity)
+        .get_mut::<ClientEntityMap>(client)
         .unwrap();
     entity_map.insert(server_entity, client_entity);
 
@@ -1183,11 +1183,8 @@ fn after_disconnect() {
     client_app.update();
     server_app.exchange_with_client(&mut client_app);
 
-    let test_client_entity = **client_app.world().resource::<TestClientEntity>();
-    server_app
-        .world_mut()
-        .entity_mut(test_client_entity)
-        .despawn();
+    let client = **client_app.world().resource::<TestClientEntity>();
+    server_app.world_mut().entity_mut(client).despawn();
     server_app.update();
 }